requirements.txt
# Python modules needed to run the B3
pymysql>=1.0.2,<2
python-dateutil>=2.8.0,<3
feedparser>=6.0.0,<7
requests>=2.25.0,<3
